+++ CHANGELOG.md
@@ -9,6 +9,8 @@ Polyverse Boost Automatic Application Modernization
 - New Performance analysis Kernel - characterizes performance of code and potential issues with 3rd party frameworks
 - New Performance quick-scan source analysis Kernel - scans code for specific issues in code and creates problems and solutions for followup
 - Added Quick Draft Architectural Blueprint Summary Kernel - based on hints about project structure, files and data
+    NOTE: Quick blueprints are only generated when a detailed Summary blueprint is not available
+- Added Project-level Analysis Command - processProject - which supports Quick Blueprint only currently
 
 ### Enhancements
 - Improved UX around file and folder analysis in Summary View in Activity Bar